refactor(ai-tools): align AI SDK adapter layout
