+++ src/ai/provider/extensions.ts
