+++ src/ai/streamManager/__tests__/withReasoningTimingMetadata.test.ts
