+++ src/ai/streamManager/withReasoningTimingMetadata.ts
