+++ src/ai/tools/adapters/aiSdk/__tests__/registry.test.ts
@@ -1,6 +1,6 @@
 import type { Tool } from 'ai';
 
-import { ToolRegistry } from '../ToolRegistry';
+import { ToolRegistry } from '../registry';
 import type { ToolApplyScope, ToolEntry } from '../types';
 
 const scope = {
