+++ src/ai/tools/adapters/aiSdk/builtin/__tests__/calendarTools.test.ts
