+++ src/ai/tools/adapters/aiSdk/builtin/__tests__/locationHealthTools.test.ts
