+++ src/ai/tools/adapters/aiSdk/builtin/deviceToolSupport.ts
