+++ src/ai/tools/adapters/aiSdk/builtin/outputSchemas.ts
