+++ src/ai/tools/adapters/aiSdk/isApprovalGated.ts
