+++ src/ai/tools/adapters/aiSdk/meta/schemaStub.ts
