+++ src/ai/tools/adapters/aiSdk/registry.ts
