+++ src/ai/tools/index.ts
@@ -1,4 +1,9 @@
-export { ToolRegistry } from './ToolRegistry';
+export { ToolRegistry } from './adapters/aiSdk/registry';
+export type {
+  RequestContext,
+  ToolApplyScope,
+  ToolDefer,
+  ToolEntry,
+} from './adapters/aiSdk/types';
 export type { ToolServiceDependencies } from './ToolService';
 export { ToolService } from './ToolService';
-export type { RequestContext, ToolApplyScope, ToolDefer, ToolEntry } from './types';
